+++ guides/foursquare.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# Copyright (C) 2014 Osmo Salomaa
+# Copyright (C) 2014 Osmo Salomaa, 2018 Rinigus
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -30,6 +30,7 @@ import poor
 import urllib.parse
 
 from concurrent.futures import ThreadPoolExecutor
+from poor.i18n import _
 
 CONF_DEFAULTS = {"sort_by_distance": False}
 
@@ -112,7 +113,9 @@ def inject_venue_details(results):
         venues = [x.response.venue for x in details]
         for i in range(len(results)):
             results[i].description = parse_description(venues[i])
+            results[i].link = parse_url(venues[i], results[i].link)
             results[i].text = parse_text(venues[i])
+            results[i].phone = parse_phone(venues[i])
 
 def nearby(query, near, radius, params):
     """Return X, Y and a list of dictionaries of places matching `query`."""
@@ -127,7 +130,10 @@ def nearby(query, near, radius, params):
     results = [poor.AttrDict(
         id=item.venue.id,
         title=item.venue.name,
+        address=parse_address(item.venue),
         description=parse_description(item.venue),
+        poi_type=parse_type(item.venue),
+        postcode=parse_postcode(item.venue),
         text=parse_text(item.venue),
         link=get_link(item.venue.id),
         x=float(item.venue.location.lng),
@@ -141,6 +147,11 @@ def nearby(query, near, radius, params):
     if results and results[0]:
         cache[url] = copy.deepcopy((x, y, results))
     return x, y, results
+
+def parse_address(venue):
+    with poor.util.silent(Exception):
+        return ", ".join(venue.location.formattedAddress).strip()
+    return ""
 
 def parse_description(venue):
     """Parse description from venue details."""
@@ -158,28 +169,36 @@ def parse_description(venue):
 def parse_text(venue):
     """Parse blurb text from venue details."""
     lines = []
-    with poor.util.silent(Exception):
-        lines.append((
-            '<font color="Theme.highlightColor">'
-            '<big>{}</big>'
-            '</font>'
-        ).format(html.escape(venue.name)))
     subtitle = []
     with poor.util.silent(Exception):
         subtitle.append((
-            '<font color="Theme.highlightColor">'
-            '<big>{:.1f}</big>'
-            '</font>'
-            '<small>&nbsp;/&nbsp;10</small>'
+            '<b>{:.1f}</b> / 10'
         ).format(venue.rating))
-    with poor.util.silent(Exception):
-        category = html.escape(venue.categories[0].name)
-        subtitle.append("<small>{}</small>".format(category))
-    lines.append("&nbsp;&nbsp;".join(subtitle))
+    lines.append("  ".join(subtitle))
     with poor.util.silent(Exception):
         tip = parse_tip(venue) or venue.get("description") or ""
         if not tip: raise ValueError("No tip")
-        lines.append("<small>{}</small>".format(html.escape(tip)))
+        lines.append("{}".format(tip))
+        lines.append("")
+    with poor.util.silent(Exception):
+        hours = venue.hours
+        open_lines = []
+        for i in hours.timeframes:
+            s = i.days + ": "
+            for j in i.open:
+                s += j.renderedTime + " "
+            open_lines.append(s.strip())
+        if len(open_lines) > 0:
+            lines.append(_("Opening hours:"))
+            lines.extend(open_lines)
+        lines.append("")
+    with poor.util.silent(Exception):
+        for i in venue.attributes.groups:
+            for j in i['items']:
+                if j.displayName!=j.displayValue:
+                    lines.append("{}: {}".format(j.displayName, j.displayValue))
+                else:
+                    lines.append(j.displayName)
     return "<br>".join(lines)
 
 def parse_tip(venue):
@@ -193,6 +212,19 @@ def parse_tip(venue):
                 if item.lang == (lang or item.lang):
                     return "“{}”".format(item.text)
 
+def parse_type(venue):
+    with poor.util.silent(Exception):
+        types=[i.name for i in venue.categories]
+        return ", ".join(types).strip()
+    return ""
+
+def parse_phone(venue):
+    with poor.util.silent(Exception):
+        return venue.contact.formattedPhone
+    with poor.util.silent(Exception):
+        return venue.contact.phone
+    return ""
+
 def prepare_point(point):
     """Return geocoded coordinates for `point`."""
     # Foursquare does geocoding too, but not that well.
@@ -201,3 +233,13 @@ def prepare_point(point):
     geocoder = poor.Geocoder("default")
     results = geocoder.geocode(point, dict(limit=1))
     return results[0]["x"], results[0]["y"]
+
+def parse_postcode(venue):
+    with poor.util.silent(Exception):
+        return venue.location.postalCode
+    return ""
+
+def parse_url(venue, default=""):
+    with poor.util.silent(Exception):
+        return venue.url
+    return default