+++ qml/GeocodePage.qml
@@ -75,17 +75,22 @@ Page {
                     // Autocompletion result with known coordinates, open directly.
                     py.call_sync("poor.app.history.add_place", [model.place]);
                     app.hideMenu();
-                    map.addPois([{
+                    var p = {
+                        "address": details.address || "",
                         "link": details.link || "",
+                        "phone": details.phone || "",
+                        "poiType": details.poi_type || "",
+                        "postcode": details.postcode || "",
                         "provider": details.provider || "",
-                        "text": details.text || details.title || model.place,
+                        "text": details.text || "",
                         "title": details.title || model.place,
                         "type": "geocode",
                         "x": details.x,
                         "y": details.y,
-                    }]);
-                    map.hidePoiBubbles();
-                    map.showPoiBubble(map.pois[map.pois.length-1]);
+                    };
+                    if (map.addPoi(p)) p = map.pois[map.pois.length-1];
+                    else p.title = app.tr("%1 [duplicate]", p.title);
+                    map.showPoi(p, true);
                     map.autoCenter = false;
                     map.setCenter(details.x, details.y);
                 } else {