+++ qml/ScaleBar.qml
@@ -51,6 +51,7 @@ Item {
             }
         }
     ]
+    visible: !app.poiActive
     width: app.navigationActive && app.portrait ? scaleBar.height : scaleBar.width
     z: 400