+++ routers/digitransit_results.qml
@@ -184,7 +184,7 @@ Page {
             onClicked: {
                 app.hideMenu();
                 map.addRoute(listItem.result);
-                map.hidePoiBubbles();
+                map.hidePoi();
                 map.fitViewToRoute();
                 map.addManeuvers(listItem.result.maneuvers);
             }