+++ routers/osmscout_results.qml
@@ -49,7 +49,7 @@ Page {
                 page.loading = false;
             } else if (route && route.x && route.x.length > 0) {
                 app.hideMenu();
-                map.hidePoiBubbles();
+                map.hidePoi();
                 map.addRoute(route);
                 map.fitViewToRoute();
                 map.addManeuvers(route.maneuvers);